Update stdin_recorder.py
I looked at your hint file and modified my code to take a similar approach when taking input from the terminal. Now this code reads in a byte at a time and writes it to the file a byte at a time as well. 
Additionally, I realize that this approach only works for Unix systems since this makes use of the `termios` library. Does this mean that the code can be designed specifically for Unix now?